requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
